requirements.txt
pandas==2.1.0
numpy==1.24.3
matplotlib==3.7.2
yfinance==0.2.28
backtrader==1.9.78.123
TA-Lib==0.11.0
schedule==1.2.0
websocket-client==1.6.3
requests==2.31.0
openpyxl==3.1.2
